Add admin layout, change naming, update header
