+++ Receptai.client/typescript/types.ts
@@ -127,10 +127,10 @@ export interface Message {
   text: string;
   show: boolean;
   label?: "Error" | "Success";
-  links?: Link[];
+  links?: MesgLink[];
 }
 
-export interface Link {
+export interface MesgLink {
   text: string;
   link: string;
   type?: "Black" | "Gray";
